evaluation/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdown>=3.5.0
aiohttp>=3.9.0
asyncio>=3.4.3
